fix: make pumpkin-pie.jpg display
When the pumpkin pie image was saved earlier, tried to change the file
extension from .webp to .jpg within the file explorer, causing the image
to be unable to display and instead showing the text from the alt
attribute.

Download a new version of the image, keeping the .webp extension, and
run it through a .jpg conversion website. Download the new valid .jpg
image and replace the invalid one. Fix the incorrect attribute href in
the img element to src.